src/repositories/text-aligner/requirements.txt
numpy
transformers
torchaudio
torch
ffmpeg-python
json
tqmd
wave